Align rotation and scale pivots to float64 (#165)
The three pivot casts used float32 while the comparable ScaleFixed path used float64. Measurement shows the asymmetry was justified rather than a defect: these three sites hand the pivot straight to _pin_source_point_under_mutation, which casts to float32 on entry, so the caller's dtype was unobservable in 0 of 200,000 slice-scale samples and bit-identical across a 128-notch train. ScaleFixed differs because it multiplies the pivot before that boundary, where float64 is closer to exact in 100% of the 55% of samples that differ.

Align them anyway so the caller stops capping precision if the rigid transform ever stores more than float32, and record the reasoning at the site. The predicted per-notch accumulation does not reproduce: 128 monotonic notches drift 0.0007 px (rotate) and 0.0011 px (scale), and the pivot the model stores stays pinned exactly.

+++ pyre/controllers/transformcontroller.py
@@ -1673,15 +1673,21 @@ class TransformController:
             edit_space = Space.Source
         model = self._TransformModel
         if isinstance(model, nornir_imageregistration.IRigidTransform):
+            # Pivots are float64 to match ScaleFixed below. These two sites hand the value
+            # straight to _pin_source_point_under_mutation, which casts to float32 on entry,
+            # so the dtype here is currently unobservable (0 of 200k slice-scale samples) --
+            # unlike ScaleFixed, which multiplies before that boundary and so needs float64
+            # to avoid double rounding. Keeping float64 means the caller stops capping
+            # precision if the rigid transform ever stores more than float32. (#165)
             if edit_space == Space.Source:
                 if center is not None:
-                    source_pivot = np.asarray(center, dtype=np.float32).ravel()[:2]
+                    source_pivot = np.asarray(center, dtype=np.float64).ravel()[:2]
                     model.RotateFixedAboutSourcePoint(rangle, source_pivot)  # type: ignore[attr-defined]
                 else:
                     model.RotateSourcePoints(rangle, None)  # type: ignore[attr-defined]
             else:
                 if center is not None:
-                    source_pivot = np.asarray(center, dtype=np.float32).ravel()[:2]
+                    source_pivot = np.asarray(center, dtype=np.float64).ravel()[:2]
                     model.RotateFixedAboutSourcePoint(rangle, source_pivot)  # type: ignore[attr-defined]
                 else:
                     model.RotateFixed(rangle, None)  # type: ignore[attr-defined]
@@ -1714,7 +1720,7 @@ class TransformController:
         model = self._TransformModel
         if isinstance(model, nornir_imageregistration.transforms.CenteredSimilarity2DTransform):
             if center is not None:
-                source_pivot = np.asarray(center, dtype=np.float32).ravel()[:2]
+                source_pivot = np.asarray(center, dtype=np.float64).ravel()[:2]
                 model.ScaleWarpedAboutSourcePoint(scale_factor, source_pivot)  # type: ignore[attr-defined]
             else:
                 model.ScaleWarped(scale_factor)  # type: ignore[attr-defined]
